Fix '+' and '-' specifity of numbers
Users now cannot spam '-' or '+' multiple times on a number.
Users are only able to use the respective operators to specify positive or negative numbers or input it as a mathSymbol.
To keep the usual "no auto change" intention, if a user presses + to specify a positive number but wants it changed to negative, the user has to press the clear button and rewrite their equation.

Delete button to delete a single character may be added in the future to ease changes to calculation.